+++ src/tests/strhashtests.cpp
@@ -70,7 +70,7 @@ void runStressTest( const std::string_view caption )
    static_assert( card < std::numeric_limits<int>::max() );
    std::array<int, card> nums{};
    std::random_device rd;
-   int sum{};
+   uint64_t sum{};
    for( int k{}; k < ntries; k++ )
    {
       T shlst;
